Initial implementation of the modular arithmetic helpers
The outer string hash function works in (mod 2**64 - 8), i.e.,
(mod 8 * (2**61 - 1)).  This modulus is nice because we can
easily normalise the result of the product of an arbitrary 64-bit
integer and a 61-bit multiplier without intermediate overflow.
It's also amenable to the usual tricks where we use the processor's
native overflow checking to drive lazy modular reduction.