Initial finalizer function
The almost-universal output of UMASH's polynomial string hash isn't
distributed quite well enough to satisfy SMHasher.  Use two steps from
the SplitMix64 update function (a xor-shift and a multiplication) to
move things around a bit.

For our collision bounds, the only thing that matters here is that
the finaliser is invertible.