Initial medium (9-16 byte) hash code
We hash 16-byte inputs normally, with a single PH iteration feeding
into the polynomial hash, but use a dedicated code path to help code
generation.  Inputs of 9-15 bytes are expanded to 16 bytes by
redundantly reading the first and last 8 bytes of data.  We avoid
extension attacks with the same length-tagging scheme that protects
UMASH in the general (long inputs) case.