Initial implementation of the long (> 16 bytes) input hash
In the general case, we compress 256-byte blocks with PH and split the
result in two 64-bit values that are fed in the toplevel polynomial
string hash.  Stop early when the last block is short, and use
redundant reads instead of zero-padding, and that's UMASH!  Just
remember to tag the last PH output with the final block's original
size in bytes to avoid extension attacks.