Initial implementation of umash_fprint
umash_fprint returns a 128-bit fingerprint with probability collision
`ceil(l / 2048)**2 * 2**-112` by concatenating two 64-bit umash
values.

Each umash_params struct describes two functions; a fingerprint is
equivalent to calling `umash_full` with `which = 0` for
`umash_fp::hash[0]`, and with `which = 1` for `umash_fp::hash[1]`.
We use a Toeplitz extension to reuse all but 32 bytes of the PH
parameter array.