umash_params_prepare: convert random bytes into safe and usable params
The outer polynomial string hash relies on having access to a fully
reduced square of the random multiplier key, for speed.  We have to
compute that value.

This also means that filling a `struct umash_params` with random bytes
gives us two extra random 64-bit words.  Use those as fallback when we
detect a 0 multiplier, or when the `PH` array includes duplicate
values.  The probability that we need more than two additional random
words are astronomically small, as long as the bytes are indeed
random.