Initial incremental hashing / fingerprinting code
The UMASH algorithm needs very little state or buffering, and only
takes the input length into account at the very end.  This makes it
an ideal candidate for incremental evaluation.

We guarantee that one can chop up a sequence of bytes in arbitrary
subsequences and, as long the subsequences are fed to incremental
umash or umash_fp in order, the result will be the same as a batch
umash or umash_fp call.