Initial implementation of umash_params_derive
Not everyone may want to hardcode a large literal struct of random
bytes.  Combine Salsa20 with `umash_params_prepare` to derive a
`struct umash_params` from a small seed and a key.  We know the odds
that `umash_params_prepare` fail are astronomically small, so we
simply increment the seed if that ever happens.  In theory, we could
thus generate the same params from different inputs.  In practice,
hardware failure is extremely more likely.