{bench,t}/exact_test: initial commit of exact testing code
Once microbenchmarking data is partitioned finely enough (e.g., group
timings by hashed data length, where any variation is external, either
from system noise or the microarchitectural state left behind by
earlier operations), we often want to answer a question of the type:

Is version A likely to be faster than version B?

This sounds like a perfect fit for the sign test... except that we
have to contend with the discrete nature of our measurements (cycle
counts), which means that identical values have positive measure.

When equal observations are not only possible, but even likely, we'll
often find ourselves in situations where a sign test says
`P(A < B) < 0.5`, and also `P(A > B) < 0.5`.  We could instead look for
`P(A <= B)`, but we'll then find `P(A <= B) > 0.5`, and `P(B >= A) > 0.5`!

Does this mean A and B are essentially indistinguishable? Hopefully not.

Once grouped by first-order parameters like input size, our
measurements aren't actually paired.  It's not incorrect to impose an
arbitrary pairing, but it's just that: arbitrary.  We may get
something more useful with an unpaired test that considers the two
input sets as a many-to-many relation.

That's what the exact test machinery introduced in this commit does.

We can define arbitrary statistics that compare two samples, and
compare that the value against the distribution of the same statistic
when the samples are shuffled pseudorandomly to sample results from
the null hypothesis.

This commit implements two statistics

- `exact_test_gt_prob` computes the sample probability that a value
  from sample A is strictly greater than a value from sample B

- `exact_test_lte_prob` computes the complementary sample probability
  that a value from smple A is less than or equal to a value from B.

We can generate a null with a balanced or imbalanced design, and
either enforce a probability that values from A are less than values
from B (? unclear if that actually works...), or apply an offset to
values from A or B in the null.  The latter offset lets us test for a
minimum effect size, e.g., "A is at least 1 less than B more
frequently than a random effect would suggest."

The python code in `t/exact_test.py` wraps the C inner loops in
`bench/exact_test.c` with a Python implementation of the Confidence
Sequence Method to perform a statistical test on the probability that
a given statistic's value was observed due to chance (under a
specified null hypothesis).

The funcion almost always (in theory) returns, with either

- `-1`: the statistic value is so low it's unlikely to have been
  generated under the null

- `1`: the statistic value is so high it's unlikely to have been
  generated under the null

- `0`: we can't say, either because we need more data, and the null
  holds exactly (the latter being somewhat improbable...).

In addition to letting ask a wider range of questions than the
Confidence Sequence Sign test, the exact test is also more powerful,
at the expense of trivially parallel CPU time for Monte Carlo
re-sampling.  For example, when comparing two quantized normals
with a center that differs by 1/10th SD,

    A = [max(0, math.ceil(random.normalvariate(1000, 10))) for _ in range(100000)]

and

    B = [max(0, math.ceil(random.normalvariate(1001, 10))) for _ in range(100000)]

the sign test with CSM needs more than 30K pairs of values
to conclude that `A < B` more than 50% of the time (p = 1e-4).

    csm.csm_driver(map(lambda a, b: a < b, A, B), 0.5, 1e-4, file=sys.stdout)
         1 1.000 0.005 1.000 -0.000
        10 0.600 0.101 0.963 -0.354
    ...
     20000 0.517 0.503 0.531 2.910
     30000 0.515 0.503 0.526 3.389
    (True, 0.5151533836904578, 0.5038949463773812, 0.5264003317024828)

The `csm_driver` code applies an internal Bonferroni correction to the
false positive rate.  However, even after downgrading the error rate
to 1e-3, we still need more than 10K pairs of values.

With the exact test, we can instead compute the (sample) percentage of
the time values from A are less-than-or-equal to value from B, and ask
how likely it is to observe a value as extreme as that.

We can get away with 1/10th as much data for the same p = 1e-4:

    exact_test.exact_test(A[:3000], B[:3000], eps=1e-4)
    actual: 0.553416
    10: 10 0 (2.397441 0.526462 / 0.553416)
    ...
    570000: 570000 0 (-12.656286 0.512436 / 0.553416)
    (1, 570000)

After 570K Monte Carlo iterations, we find that the sample probability
that values from A <= values from B (55.3%) is greater than what one
would expect if the two underlying distributions were identical.  This
also happens to be the minimum number of iterations needed to conclude
anything at p = 1e-4, with our implementation of the exact test: that's
how many predicate values the CSM needs to determine that the success
rate differs from 1e-4 with low enough error rate.

Of course, in order to only sample 3000 data points from each
distribution, we had to use a lot more CPU cycles during the analysis:
570K pseudo-random resamplings (each at least linear-time in the
number of data points), instead of 30000 constant-time comparisons.

Sometimes, we won't have enough data to reach a conclusion.  The exact
test also recognises this situation in finite time.

TESTED=manual smoke test.