bench/exact_test: use a hybrid counting / comparison sort
We expect our discrete data to have a dense region with a lot of
repeated values, and a few outliers.  Optimise the sort algorithm for
this sort of input: detect when the input is dense (or find a dense
subrange), apply a counting sort for that dense region and handle the
remaining ouliers with a call to qsort(3).  We can merge the outliers
in the same function that converts the occurrence histogram for the
dense range into repeated sorted values.

The bottleneck in the exact permutation test loop is now the
Fisher-Yates shuffle.

Whenever the `bench.so` library is loaded by `umash.py` (whenever
`UMASH_BENCH_LIB=[path o bench.so]` is set), `t/run-tests.sh` will run
the Python code in `test_bench_exact_test.py`.  When that environment
variable is left unset, the test is skipped.

TESTED=new conditionally-enabled Python test.