bench/exact_test: add more statistic functions
It's nice that we can easily compare a lot of quantiles without
needing needing a ton of data, but the real strength of permutation
testing is that we can compare arbitrary statistics.

Implement a statistic that estimates the probability that a random
value from A yields a value less than or equal to a random value from
B (lte_prob), and another for the complementary probability that a
random value from A is strictly greater than a random value from B
(gt_prob).  When `lte_prob > 0.5`, this essentially tells us that
switching from B to A would be a win or a wash in the majority of
cases.  With equality, it is possible for lte_prob(x, y) > 0.5,
and lte_prob(x, y) > 0.5 as well.

Also add a truncated mean statistic: the mean is a useful metric,
but is easily swayed by ouliers.  We simply compute the mean after
removing a fixed fraction of the most extreme data points.

TESTED=new python tests with reference implementations.