t/exact_test: parallelise with a process pool
We use a weird scheme with futures that never return and instead write
to a shmem queue because we don't want to waste too much time pickling
a large data set in the arguments, but wish to receive the results ASAP,
to exit early when possible.

The next commit will join on these futures, so it'll be *less* weird.

TESTED=existing tests.