t/exact_test_sampler: switch to a simpler Future based dataflow
Get rid of the queue: while useful, it doesn't really scale well
to RPC-like designs.

Switch to regular Pool AsyncResults, with a primary trigger to return
values based on real time (as a fraction of the total time spent in
the computation so far), and a secondary trigger on total number of
values, mostly to avoid returning extremely large results.

TESTED=existing tests.