t/exact_test_sampler_server: clamp auto-scaling
Yeah, we want to handle more concurrent requests when we have more
cores, but at some point, it stops making sense: we're still stuck on
the GIL.  Arbitrary limit the number of concurrent requests to 32,
because that seems more reasonable than the 512 the previous code
attempted to handle on a 256-thread Xeon Phi.

TESTED=ran the server.