umash: change the finalizer to xor-rot
The low 32 bits are fine with just xor-shift.  However, we need to mix
in the high bit some more, if we want to get rid of the
multiplication.  Rotations would be perfect, but plain xor-rot isn't
invertible (e.g., 0 and -1 both map to 0).

Marc B. Reynolds shows how [xor-2-rotate map to invertible matrices](https://marc-b-reynolds.github.io/math/2017/10/13/XorRotate.html),
so we use that.  Xoring after bit rotations (instead of right shifts)
affects both the high and low bits, and xoring two different rotations
guarantees invertibility.

In terms of latency, we replaced one multiplication on the critical
path with a xor.  This improves end-to-end latency for 9-16 byte
inputs by ~1.5 cycles/call on average, on a Kaby Lake laptop.  It's
also an improvement over the previous code more than 60% of the time.

Why 8 and 33? We evaluated all (63 * 62 / 2) options exhaustively, and
this was one of the options for which SMHasher reported the least
worst-case bias for the test we looked at.  The rotation amounts are
also SIMD-friendly: 8 is a byte shuffles, and 33 = 32 + 1 is a one-bit
rotation followed by a byte/dword shuffle.

TESTED=reference code. also test that the finalizer can be inverted.