umash: switch to hybrid "OH" block compressor
NH is good for latency, awful on throughput. PH is good on throughput,
not great on latency.  Use a hybrid "OH" compressor: PH for all but
the last 16-byte chunk, and NH for that last chunk.

We need a bit more elbow grease for the universality proofs, but
nothing heroic.

The result is improved latency for short input hash, and a wash on
short input fingerprint.  Longer inputs (more than 256 bytes) can be a
bit slower, but SMHasher shows a marginal throughput reduction (on the
Xeon 8175M) or improvement (on a Kaby Lake laptop), so any downside on
just-large-enough inputs seems reasonable.

The new algorithm also has less symmetries; there may new low-hanging
fruits to pick at.

TESTED=updated tests.