NFC umash_reference: reorganize OH compression to follow blog
http://pvk.ca/Blog/2020/10/31/nearly-double-the-ph-bits-with-one-more-clmul/
uses a nicer presentation in terms of abstract mixing functions,
and combining those cleverly.

Make the reference code do the same thing; we want to reverse the
order because the twisting logic for the secondary hash is simpler for
lower indexes, but we want the real implementation to make the *last*
chunks simpler (for latency).