umash: switch secondary hash to twisted version
described in http://pvk.ca/Blog/2020/10/31/nearly-double-the-ph-bits-with-one-more-clmul/

It's now slightly more work to compute the secondary hash by itself,
but almost half as much work when we compute a full fingerprint.

This already helps short input latency, but throughput improvements
aren't in this commit.  A hackily souped up version did confirm that
we can at least match the current throughput, and it seems reasonable
to think the new algorithm is better able to use 256- or 512- bit
vectors.