umash: get rid of the full-block OH compressor
Turns out it's faster to always use the variable-length compressor on
a EPYC 7713... and we'll add *really* specialised code for long input
soon.  In the meantime, less code and I$ pressure is a definite win.

Unclear on what the effect comes from, except maybe I$ pressure. It
looks like gcc prefers to implement the outer hashing while loop
as the usually `if / do { } while`, so I'm not seeing any obvious
unlocked cleverness here.

Closes #19.

TESTED=unit tests, and performance testing notebooks.