umash_long.inc: add runtime feature dispatch logic
When UMASH_DYNAMIC_DISPATCH is true (the default for x86-64), the
first call to `umash_multiple_blocks` figures out the appropriate
implementation for the current CPU, stashes the result, and calls it.

Subsequent calls to `umash_multiple_blocks` will directly jump to the
implementation picked by the first call.

Benchmarking shows a ~30 cycle slowdown for 64K (compared to ~8600
cycles), i.e. < 0.4%.  Seems worth it for the much greater speed-up
we can get from a specialised implementation.

In preparation for #21's AVX2 / VPCLMULQDQ backend.

TESTED=new and existing tests.