t/umash_substring.c: new test program
This new test program expects a number of bytes in argv[1], reads that
many bytes from stdin, and:

1. prints throughput estimates for hashing that many bytes
2. computes and prints hash values for substrings of the input,
   at various alignments.

The second phase is useful to confirm that we avoid instructions that
assume alignment, and to integrate directly with tools like valgrind.
Comparing the printed output also gives us more confidence that
different implementations of UMASH compute the same function.

TESTED=ran the program.