umash_long: use partially reduced accumulator for the Horner poly
The double-pumped Horner update starts with `m0 * (acc + h0)`.
Representing `acc` as an evaluated sum of `base + 8 * fixup` where
fixup is small (at most 3) makes it easy to implement `h0 + 8 * fixup`
with a regular overflow-oblivious addition (e.g., `LEA`): `h0` is a
~uniformly distributed u64, so we can branch in the unlikely care that
`h0` is so large that `h0 + 8 * fixup` might overflow.

This lazier representation for the accumulator lets us add the two
multiplicands (`m0 * (acc + h0)` and `m1 * h1`) with fewer
instructions, and folding `fixup` in `h0` is nearly free.

This change is performance neutral for 64 KB inputs on a EPYC 7713.

TESTED=new unit test.