umash_long: add a fingerprint function for multiple blocks
and divert fingerprinting calls to it.

For now, the new function is just a pure copy of the regular code
path: let's start by adding test coverage before we try to improve
performance.  Just having the new function apparently helps lower
latency by 8% on 64 KB inputs.

There is other no statistically significant impact on hashes <= 16
bytes, and small effects in both directions for inputs 16-640 bytes.
Slowdowns are on the order of 1-2 cycles.  Seems worth it for the
ability to dispatch to a specialised fingerprinting function for long
inputs.

Part of #21 for 128-bit fingerprints.

TESTED=new and existing tests.