umash_long.inc: simplify initialisation of `lrc_init`
in `umash_fprint_multiple_blocks_vpclmulqdq`.  This only happens
once in the function, but we might as well get it right.

GCC likes to emit a `MOVQ` + `PINSERT` when populating the low half of
`_mm256_set_epi64x`.  We can simply use an unaligned SSE load.

No impact on performance, as expected.

TESTED=existing tests.