README: update performance numbers
With baremetal access to an EPYC 7713, it's now possible to report
useful cycle counts, in addition to throughput numbers.

The bulk hashing and fingerprint performance work also brings us a lot
closer (if not over) to the throughput of other contemporary
hardware-optimised hashes.  In fact, the fingerprinting function is
now in the same league as murmurhash3 and farmhash.

TESTED=readme only.