umash_long: simplify generic bulk fprint function
When fingerprinting large inputs, it's trivial to determine when we're
processing the last PH chunk in a block: the loop is fully unrolled.

We can thus simplify the logic for the "shifted" PH accumulator,
and save a register.  This improves throughput for the generic
code path on an EPYC 7713 by ~3.5% (13120 -> 12680 cycles).

```
[(65536,
  {'mean': Result(actual_value=-524.5595595595596, judgement=-1, m=20000, n=20000, num_trials=2750000),
   'lte': Result(actual_value=0.803463475, judgement=1, m=20000, n=20000, num_trials=2750000),
   'q99': Result(actual_value=-2520.0, judgement=-1, m=20000, n=20000, num_trials=2750000),
   'q99_sa': Result(actual_value=-2520.0, judgement=-1, m=20000, n=20000, num_trials=2750000)})]
```

TESTED=existing tests, on a machine that does not have VPCLMULQDQ.