smhasher: update test logs
On a 2.0 GHz EPYC 7713 with boost enabled.  The 128-bit and
primary (umash64/umash32/umash32_hi) hash timings are on an
unloaded box.

The secondary hash's performance numbers aren't that relevant:
it rarely makes sense to use it by itself.  However, we do want
to test it for obvious badness, since it's far from identical to
the primary hash.

TESTED=n/a