umash_long: probe for OS support before using AVX
Even when the chip supports VPCLMULQDQ, we don't want to use the
high half of AVX registers when the OS won't save them correctly
on context switches.  AVX has been around for more than 10 years,
so rare is the kernel that does not handle it correctly, but it
doesn't hurt to get it right.