umash_long: optionally cache all PH parameters in registers
We need 15x 128-bit registers in `umash_multiple_blocks_generic` if we
want to cache all PH parameters.  That's not an option with regular
SSE/AVX, but is reasonable for AVX-512 (useful because using the full
32 vector register set does *not* increase power consumption) or on
aarch64 neon.

TESTED=in SSE mode with overridden values for UMASH_LONG_PH_PARAMS_IN_REGS.