Fix array size warnings
In two places, byte arrays are populated using string constants that
include a NUL terminator, but the array sizes do not include space for
the terminator. This generates warnings on recent compilers. Convert the
string constants to character arrays to quench these warnings.

It would also be possible to use __attribute__((nonstring)) to silence
the warnings, but this attribute doesn't exist on older compilers so
extra checks would be needed to determine that this attribute is present.